Rename mio header
R CMD check strangely warns about the .hpp extension
